requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.2
lxml==4.9.3
pyvis==0.3.2
networkx==3.2.1
Jinja2>=3.1.3
streamlit==1.38.0
